requirements.txt
testrail-api>=1.13.4
typer[all]>=0.9.0
rich>=13.0.0
pyyaml>=6.0
requests>=2.31.0
